renamed instructor programs to course
