+++ src/components/pages/InstructorAddCourse/InstructorAddedCoursesList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import InstructorProgram from './InstructorProgram';
+import InstructorProgram from './InstructorAddCourse';
 import { connect } from 'react-redux';
 import { List } from 'antd';
 
@@ -8,8 +8,8 @@ const InstructorProgramList = props => {
 
   return (
     <div>
-      <h1>Already Submitted Programs:</h1>
-      <List className="instructor-booking-card">
+      <h1 className="instructor-h1">Already Submitted Programs:</h1>
+      <List className="instructor-list">
         {programs.own_programs &&
           programs.own_programs.map(program => {
             return <InstructorProgram key={program.id} program={program} />;
@@ -24,14 +24,3 @@ const mapStateToProps = state => {
 };
 
 export default connect(mapStateToProps, {})(InstructorProgramList);
-
-//styles for the card original
-// style={{ display: 'flex', justifyContent: 'center' }}
-
-//style for h1
-// style={{
-//   marginBottom: 10,
-//   fontSize: 20,
-//   display: 'flex',
-//   justifyContent: 'center',
-// }}
