+++ src/styles/InstructorStyles/index.less
@@ -127,6 +127,7 @@
 .instructor-booking-card {
   display: flex;
   flex-direction: row;
+  justify-content: space-between;
   background: white;
   margin: 1%;
   border-radius: 0.25rem;
@@ -194,4 +195,17 @@
 .ant-layout-sider-trigger {
   position: relative !important;
   z-index: 0 !important;
+}
+
+//instructor add course style
+.instructor-h1 {
+  margin-bottom: 10;
+  font-size: 20;
+  display: flex;
+  justify-content: center;
+}
+
+.instructor-list {
+  display: flex;
+  justify-content: center;
 }
